Add visual testing to retro page cypress tests
