+++ ui/cypress/support/index.d.ts
@@ -15,6 +15,7 @@
  * limitations under the License.
  */
 import 'cypress-jest-adapter';
+import '@types/cypress-image-snapshot';
 
 import Topic from '../../src/types/Topic';
 
@@ -49,10 +50,10 @@ declare global {
 
 			shouldBeOnRetroPage(teamId: string): Chainable<void>;
 
+			shouldCreateActionItems(actionItems: string[]): Chainable<void>;
+
 			switchToDarkMode(): Chainable<void>;
 			switchToLightMode(): Chainable<void>;
-
-			matchImageSnapshot(options?: any): Chainable<void>;
 		}
 	}
 }
