+++ ui/cypress/tsconfig.json
@@ -18,7 +18,7 @@
   "compilerOptions": {
     "target": "es5",
     "lib": ["es5", "dom"],
-	"types": ["cypress", "@testing-library/cypress"]
+	"types": ["cypress", "@testing-library/cypress", "@types/cypress-image-snapshot"]
   },
   "include": ["**/*.ts"]
 }
