+++ ui/package.json
@@ -72,6 +72,7 @@
     "@testing-library/react": "^12.1.2",
     "@testing-library/react-hooks": "^7.0.2",
     "@testing-library/user-event": "^13.1.1",
+    "@types/cypress-image-snapshot": "^3.1.6",
     "@types/file-saver": "^2.0.5",
     "@types/hammerjs": "^2.0.41",
     "@types/jest": "^27.4.0",
